moved song creation to profile page
